refactory sprint planning

+++ src/modules/sprint-planning/application/pages/useController.tsx
@@ -1,29 +1,40 @@
 import { useEffect, useState } from "react";
 
 import { useService } from "../../../../di/DecouplerContext";
-import AlertModalService from "../../../core/components/AlertModal/AlertModalService";
-import AppModalService from "../../../core/components/AppModal/AppModalService";
-import { IActions, IColumns } from "../../../core/components/DisplayTable";
-import { IHeaderActions } from "../../../core/components/DisplayTable/headerActions";
+import AlertModalService, {
+  modalType,
+} from "../../../core/components/AlertModal/AlertModalService";
+import { IActions, IColumns } from "../../../core/components/DynamicTable";
+import { IHeaderActions } from "../../../core/components/DynamicTable/headerActions";
 import Task from "../../domain/Task";
 import TaskRepository from "../../repository/TaskRepository";
 import TaskModel from "../data/TaskModel";
-import TaskStore from "../data/TaskStore";
-import TaskForm from "./form";
+import { BackupService } from "../../../../services/BackupService";
 
 const useController = () => {
-  const modalService = useService<AppModalService>("AppModalService");
   const alertService = useService<AlertModalService>("AlertModalService");
-  const taskStore = useService<TaskStore>("TaskStore");
   const taskRepository = useService<TaskRepository>("TaskRepository");
   const [lines, setLines] = useState<TaskModel[]>([]);
+  const [emptyLine, setEmptyLine] = useState(new TaskModel());
+
   const tColumns: IColumns[] = [
-    { field: "parentId", title: "Nº item pai" },
-    { field: "externalId", title: "Nº item" },
-    { field: "title", title: "Título" },
-    { field: "descriptionTruncated", title: "Descrição" },
-    { field: "owner", title: "Responsável" },
-    { field: "sprint", title: "Sprint" },
+    { field: "parentId", title: "Nº item pai", width: 110 },
+    { field: "taskId", title: "Nº item", width: 110 },
+    { field: "type", title: "Tipo", width: 100, datalist: ["Story", "Task", "Bug", "Subtask"] },
+    { field: "description", title: "Descrição", required: true, indent: true },
+    { field: "owner", title: "Responsável", width: 150 },
+    {
+      field: "sprint",
+      title: "Sprint",
+      width: 60,
+      datalist: ["SP1", "SP2", "SP3", "SP4", "SP5", "SP6"],
+    },
+    {
+      field: "status",
+      title: "Status",
+      width: 100,
+      datalist: ["Backlog", "A fazer", "Em andamento", "Concluída"],
+    },
   ];
 
   useEffect(() => {
@@ -33,47 +44,73 @@ const useController = () => {
   useEffect(() => {
     taskRepository.getAll();
     var subscriber = taskRepository.data$.subscribe(items => {
-      const _max = 20;
-      const _description = (d?: string) =>
-        (d ?? "").length > _max ? `${d?.substring(0, _max)}...` : d ? d : "-";
-      setLines(
-        items.map(item => ({
+      const newEmpty = new TaskModel();
+      setEmptyLine(newEmpty);
+      if (items.length === 0) {
+        setLines([newEmpty]);
+        return;
+      }
+      setLines([
+        ...items.map(item => ({
           ...TaskModel.fromDomain(item),
-          descriptionTruncated: _description(
-            TaskModel.fromDomain(item).description,
-          ),
+          description:
+            item.parentId && item.type?.toLowerCase().indexOf("sub") !== -1
+              ? `\t\t${item.description}`
+              : item.description,
         })),
-      );
+        newEmpty,
+      ]);
     });
     return () => {
       subscriber.unsubscribe();
     };
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [taskRepository]);
+
+  const handleFieldChange = (line: TaskModel, field: any, value: any) => {
+    const linesUpdated = lines.map(curLine =>
+      curLine.id === line.id ? { ...curLine, [field]: value } : curLine,
+    );
+    if (line.id === emptyLine.id) {
+      const newEmpty = new TaskModel();
+      setEmptyLine(newEmpty);
+      linesUpdated.push(newEmpty);
+    }
+    setLines(linesUpdated);
+  };
 
   const handleSave = () => {
     try {
-      const model = taskStore.current;
-      const task = Task.create(
-        model.title,
-        model.description,
-        model.externalId,
-        model.owner,
-        model.sprint,
-        model.parentId,
-      );
-      if (!model.id) taskRepository.create(task);
-      else taskRepository.update(model.id, task.updateId(model.id));
-      modalService.close();
+      taskRepository.clear();
+      lines.forEach(line => {
+        const model = line;
+        if (!model.description) return;
+        const task = Task.create(
+          model.description?.trim(),
+          model.parentId,
+          model.taskId,
+          model.type,
+          model.owner,
+          model.sprint,
+          model.status,
+        );
+        task.updateId(model.id);
+        taskRepository.create(task);
+      });
+      showMessage("success", "Salvo com sucesso", "Aviso");
+      setTimeout(() => {
+        alertService.close();
+      }, 1000);
     } catch (e: any) {
       showMessage("error", e.message);
     }
   };
 
-  const showMessage = (type: "error" | "info", message: string) => {
+  const showMessage = (type: modalType, message: string, title: string = "Erro") => {
     alertService
       .config({
         type: type,
-        title: "Erro",
+        title: title,
         buttonOkLabel: "Ok",
         buttonCancelHidden: true,
         children: () => <span>{message}</span>,
@@ -81,51 +118,14 @@ const useController = () => {
       .open();
   };
 
-  const handleNew = () => {
-    const model = new TaskModel();
-    taskStore.updateCurrent(model);
-    modalService
-      .config({
-        title: "novo item",
-        size: "large",
-        buttonOkLabel: "Criar",
-        buttonOkAction: handleSave,
-        children: () => (
-          <TaskForm
-            data={model}
-            onChange={state => {
-              taskStore.updateCurrent(state);
-            }}
-          />
-        ),
-      })
-      .open();
-  };
+  const handleImport = () => {};
 
-  const handleEdit = (line: TaskModel) => {
-    const model = lines.find(x => x.id === line.id);
-    if (!model) return;
-    modalService
-      .config({
-        title: `editar item (${line.id.split("-")[0]})`,
-        size: "large",
-        buttonOkLabel: "Salvar",
-        buttonOkAction: handleSave,
-        children: () => (
-          <TaskForm
-            data={model}
-            onChange={state => {
-              taskStore.updateCurrent(state);
-            }}
-          />
-        ),
-      })
-      .open();
+  const handleExport = () => {
+    BackupService.exportDataAsCsv(lines);
   };
 
   const handleDelete = (line: TaskModel) => {
-    if (window.confirm("Excluir item?"))
-      taskRepository.delete(line.id);
+    if (window.confirm("Excluir item?")) taskRepository.delete(line.id);
   };
 
   const tActions: IActions[] = [
@@ -136,13 +136,21 @@ const useController = () => {
   ];
 
   const tHeaderButtons: IHeaderActions = {
-    buttonNew: {
-      label: "Novo",
-      action: handleNew,
+    buttonSave: {
+      label: "Salvar",
+      action: handleSave,
+    },
+    buttonImport: {
+      label: "Importar",
+      action: handleImport,
+    },
+    buttonExport: {
+      label: "Exportar",
+      action: handleExport,
     },
   };
 
-  return { tActions, tColumns, tHeaderButtons, lines, handleNew, handleEdit };
+  return { tActions, tColumns, tHeaderButtons, lines, handleFieldChange };
 };
 
 export default useController;
